Cast wsize as early as possible in slide_hash for AVX2 and SSE2.

+++ arch/x86/slide_sse.c
@@ -16,8 +16,8 @@
 ZLIB_INTERNAL void slide_hash_sse2(deflate_state *s) {
     Pos *p;
     unsigned n;
-    unsigned wsize = s->w_size;
-    const __m128i xmm_wsize = _mm_set1_epi16((uint16_t)wsize);
+    uint16_t wsize = (uint16_t)s->w_size;
+    const __m128i xmm_wsize = _mm_set1_epi16(wsize);
 
     n = s->hash_size;
     p = &s->head[n] - 8;
